requirements.txt
Flask==2.3.3
tensorflow
pillow
numpy